move pagemenuedit to formedit directory and do some cleanup
